+++ src/main/kotlin/ru/my/test/entity/Contact.kt
@@ -5,7 +5,7 @@ import javax.persistence.*
 @Entity
 @Table(name = "contacts")
 class Contact(
-    id: Int = 0,
+    id: Long = 0,
     var phone: String,
     var email: String,
 ) : BaseEntity(id) {
