+++ src/main/kotlin/ru/my/test/service/BookRepository.kt
@@ -4,17 +4,17 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ru.my.test.entity.Book
 import ru.my.test.model.NotFoundException
 
-interface BookRepository : JpaRepository<Book, Int> {
+interface BookRepository : JpaRepository<Book, Long> {
     fun existsByName(name: String): Boolean
 }
 
-fun BookRepository.findOrException(id: Int): Book {
+fun BookRepository.findOrException(id: Long): Book {
     return this.findById(id).orElseThrow {
         NotFoundException("Не удалось найти книгу с id: $id")
     }
 }
 
-fun BookRepository.findAllByIdOrException(booksIds: List<Int>): List<Book> {
+fun BookRepository.findAllByIdOrException(booksIds: List<Long>): List<Book> {
     val books = this.findAllById(booksIds)
 
     booksIds.forEach { bookId ->
