+++ src/main/kotlin/ru/my/test/service/ContactRepository.kt
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ru.my.test.entity.Contact
 import java.util.*
 
-interface ContactRepository : JpaRepository<Contact, Int>
+interface ContactRepository : JpaRepository<Contact, Long>
 
-fun ContactRepository.findOrException(id: Int): Contact {
+fun ContactRepository.findOrException(id: Long): Contact {
     return this.findById(id).orElseThrow {
         NotFoundException("Не удалось найти контактные данные с id: $id")
     }
